+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ![](https://img.shields.io/badge/release%20date-November%2015%2C%202025-blue)
 ![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.5.4/module.zip)
 
-- In certain configurations, an error was being through during migration from v1.5.0.  Now fixed
+- In certain configurations, an error was preventing migration from v1.5.0 to v1.5.1 or later.  Now fixed.
 
 ## v1.5.3 - Fixed bug adding journal entries to lore
 ![](https://img.shields.io/badge/release%20date-November%2011%2C%202025-blue)