+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## v1.5.5 - Bug fixes 
+![](https://img.shields.io/badge/release%20date-November%2021%2C%202025-blue)
+![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.5.5/module.zip)
+
+- You can now correctly create new types in the type selection box on the entry description tab
+
+
 ## v1.5.4 - Fixed migration bug 
 ![](https://img.shields.io/badge/release%20date-November%2015%2C%202025-blue)
 ![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.5.4/module.zip)